MALI: add gpuinfo sysfs entry
Add gpuinfo sysfs read-only entry to get a string with the Mali
GPU product name, the number of cores, the hardware revision and
the raw product id.  For example, on a Chromebook (Exynos 5250):

    Mali-T60x MP4 r0p0 0x6956

If the GPU product name is unknown, "(Unknown Mali GPU)" will be
shown followed by the same hardware information.  The raw product
id can then be used to identify the GPU via other means.

Signed-off-by: Guillaume Tucker <guillaume.tucker@arm.com>